PowerShell CopyAll Script
This script is designed to take an array of directories and copy them to
the supplied target directory. There will likely also be a parameter to
clone the original directory structure during the copy.

It cam about as I had a number of unrelated directories that needed to
be copied to the same drive.

Still in development.