PowerShell RenameMP3 Script
This script is designed to take a directory containing MP3 files and
rename them to my preferred standard using the information in the MP3
Tag.

It came about as I currently have over 18000 MP3 files which are not all
well named or at least not named to my standard.

Still in development.