PowerShell RemoveAll Script
This script is designed to take an array of directories/files and
permanently deletes them.

It came about as my custom backup scripts purposely leave the old files
and directories in place until they are deemed safe to remove. This is,
unfortunately, a manual process of finding the files and deleting them
or "quickly" throwing together a bunch of del commands and running them
in sequence to get rid of them. This script expands on the del commands
idea and makes it more extensible.

Still in development.